continued clean up

+++ src/main/scala/com/ldaniels528/robowars/events/MotionCommand.scala
@@ -6,4 +6,4 @@ import com.ldaniels528.fxcore3d.{FxEvent, FxPoint3D}
  * Represents a motion event
  * @author lawrence.daniels@gmail.com
  */
-case class MotionCommand(time: Double, id: Int, command: Int, dest: FxPoint3D, precision: Double) extends FxEvent
+case class MotionCommand(time: Double, id: Long, command: Int, dest: FxPoint3D, precision: Double) extends FxEvent
